+++ src/sections/notes/components/NotesEditor/useNotesDebounceUpdate.ts
@@ -3,103 +3,82 @@ import { useEffect, useRef } from 'react';
 import type { NoteDTO } from '@api/notes';
 
 interface UseNotesDebounceUpdateProps {
-  note: NoteDTO | null;
+  selectedNote: NoteDTO | null;
   onUpdate: (id: number, title: string | undefined, body: string | undefined) => Promise<void>;
 }
 
-export const useNotesDebounceUpdate = ({ note, onUpdate }: UseNotesDebounceUpdateProps) => {
+export const useNotesDebounceUpdate = ({ selectedNote, onUpdate }: UseNotesDebounceUpdateProps) => {
   const timerRef = useRef<NodeJS.Timeout | null>(null);
-  const lastValuesRef = useRef({
-    noteId: note?.id,
-    title: note?.title,
-    body: note?.body,
-  });
-
-  const noteId = note?.id;
-  const noteTitle = note?.title;
-  const noteBody = note?.body;
+  const editingNoteId = useRef(selectedNote?.id);
+  const pendingChangesRef = useRef<{
+    title?: string;
+    body?: string;
+  }>({});
+  const selectedNoteId = selectedNote?.id;
 
   // Update tracking ref on the note change
   useEffect(() => {
-    if (lastValuesRef.current.noteId !== noteId) {
+    if (editingNoteId.current !== selectedNoteId) {
       // reset ref values if selected note changed
-      lastValuesRef.current = {
-        noteId: noteId,
-        title: noteTitle,
-        body: noteBody,
-      };
-    } else if (!timerRef.current) {
-      // sync with props if there is no pending update
-      lastValuesRef.current.title = noteTitle;
-      lastValuesRef.current.body = noteBody;
+      editingNoteId.current = selectedNoteId;
+      pendingChangesRef.current = {};
     }
-  }, [noteId, noteTitle, noteBody]);
+  }, [selectedNoteId]);
 
   // Use debounce timer to reduce number of update api requests
   // send request only when user stopped typing for 1000ms
   const debounceUpdate = (newTitle: string | undefined, newBody: string | undefined) => {
-    let changed = false;
-
-    if (newTitle !== undefined && newTitle !== lastValuesRef.current.title) {
-      lastValuesRef.current.title = newTitle;
-      lastValuesRef.current.body = undefined;
-      changed = true;
-    } else if (newBody !== undefined && newBody !== lastValuesRef.current.body) {
-      lastValuesRef.current.body = newBody;
-      lastValuesRef.current.title = undefined;
-      changed = true;
+    if (!selectedNoteId) {
+      return;
     }
 
-    if (note && (changed || timerRef.current)) {
-      if (timerRef.current) {
-        clearTimeout(timerRef.current);
-      }
+    if (newTitle !== undefined && newTitle !== selectedNote?.title) {
+      pendingChangesRef.current.title = newTitle;
+    }
+    if (newBody !== undefined && newBody !== selectedNote?.body) {
+      pendingChangesRef.current.body = newBody;
+    }
 
+    // clear timer on every user input
+    if (timerRef.current) {
+      clearTimeout(timerRef.current);
+      timerRef.current = null;
+    }
+
+    // in case of actual changes, start timer
+    if (pendingChangesRef.current) {
       timerRef.current = setTimeout(() => {
-        const {
-          noteId: currentNoteId,
-          title: currentTitle,
-          body: currentBody,
-        } = lastValuesRef.current;
-
         // Ensure we are still editing the same note
-        if (currentNoteId === note.id) {
-          void onUpdate(note.id, currentTitle, currentBody);
+        if (editingNoteId.current === selectedNote.id) {
+          const { title: currentTitle, body: currentBody } = pendingChangesRef.current;
+          void onUpdate(selectedNote.id, currentTitle, currentBody);
+          pendingChangesRef.current = {};
         }
+
         timerRef.current = null;
       }, 1000);
     }
   };
 
-  // If there is a pending update when the selected note changes or component unmounts,
-  // execute it immediately to ensure no data is lost.
+  // unmount cleanup
   useEffect(() => {
     return () => {
-      if (timerRef.current && note) {
-        clearTimeout(timerRef.current);
-        const {
-          noteId: currentNoteId,
-          title: currentTitle,
-          body: currentBody,
-        } = lastValuesRef.current;
-
-        // Ensure we are still editing the same note
-        if (currentNoteId === note.id) {
-          void onUpdate(note.id, currentTitle, currentBody);
-        }
-        timerRef.current = null;
-      }
-    };
-  }, [noteId, note, onUpdate]);
-
-  return {
-    debounceUpdate,
-    hasPendingUpdate: !!timerRef.current,
-    clearDebounceTimer: () => {
+      // clear timer
       if (timerRef.current) {
         clearTimeout(timerRef.current);
         timerRef.current = null;
       }
-    },
+
+      // save pending changes for editing note
+      if (selectedNoteId && editingNoteId.current === selectedNoteId && pendingChangesRef.current) {
+        const { title: pendingTitle, body: pendingBody } = pendingChangesRef.current;
+        void onUpdate(selectedNoteId, pendingTitle, pendingBody);
+      }
+    };
+  }, [selectedNoteId, onUpdate]);
+
+  return {
+    debounceUpdate,
+    hasPendingUpdate: !!timerRef.current,
   };
 };