+++ src/sections/settings/index.ts
@@ -1,0 +1,4 @@
+export * from './Settings';
+export * from './components/SettingsProfile/SettingsProfile';
+export * from './components/SettingsPlanner/SettingsPlanner';
+export * from './components/SettingsNotes/SettingsNotes';